Separating WorkspaceDataCapacityService from WorkspaceService
